# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nixyah API - Food marketplace for Côte d'Ivoire
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface RegisterClientInput {
  name: string;
  email?: string;
  phone?: string;
  password: string;
  location: string;
  preferences?: string[];
}

export interface RegisterChefInput {
  name: string;
  email?: string;
  phone?: string;
  password: string;
  specialty: string;
  location: string;
  zone: string;
  bio: string;
  priceRange: string;
  coverColor?: string;
  specialties?: string[];
}

export interface LoginInput {
  emailOrPhone: string;
  password: string;
}

export type UserProfileType =
  (typeof UserProfileType)[keyof typeof UserProfileType];

export const UserProfileType = {
  client: "client",
  chef: "chef",
} as const;

export interface Dish {
  id: string;
  chefId: string;
  name: string;
  description: string;
  price: number;
  category: string;
  prepTime: string;
  isPopular?: boolean;
}

export interface Story {
  id: string;
  chefId: string;
  chefName: string;
  chefCoverColor: string;
  caption: string;
  dishName?: string;
  price?: number;
  emoji?: string;
  bgColor?: string;
  createdAt: string;
  expiresAt: string;
}

export interface ChefProfile {
  id: string;
  userId: string;
  name: string;
  specialty: string;
  location: string;
  zone: string;
  bio: string;
  rating: number;
  reviewCount: number;
  priceRange: string;
  isVerified: boolean;
  isOnline: boolean;
  coverColor: string;
  responseTime: string;
  dishes?: Dish[];
  stories?: Story[];
}

export interface UserProfile {
  id: string;
  name: string;
  email?: string;
  phone?: string;
  type: UserProfileType;
  location: string;
  coverColor?: string;
  chefProfile?: ChefProfile;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface CreateStoryInput {
  caption: string;
  dishName?: string;
  price?: number;
  emoji?: string;
  bgColor?: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  accepted: "accepted",
  preparing: "preparing",
  ready: "ready",
  delivered: "delivered",
} as const;

export interface OrderItem {
  dishId: string;
  dishName: string;
  quantity: number;
  price: number;
}

export interface Order {
  id: string;
  clientId: string;
  chefId: string;
  chefName: string;
  status: OrderStatus;
  total: number;
  occasion?: string;
  persons?: number;
  notes?: string;
  createdAt: string;
  items?: OrderItem[];
}

export interface CreateOrderInput {
  chefId: string;
  occasion?: string;
  persons?: number;
  budget?: string;
  notes?: string;
  items?: OrderItem[];
}

export interface Chat {
  id: string;
  chefId: string;
  chefName: string;
  chefSpecialty: string;
  chefCoverColor: string;
  lastMessage?: string;
  lastMessageAt?: string;
  unreadCount: number;
}

export interface Message {
  id: string;
  chatId: string;
  senderId: string;
  text: string;
  isMe: boolean;
  createdAt: string;
}

export interface SendMessageInput {
  text: string;
  chefId?: string;
}

export type ListChefsParams = {
  specialty?: string;
  location?: string;
  online?: boolean;
};

export type ListChefs200 = {
  chefs: ChefProfile[];
};

export type ListStories200 = {
  stories: Story[];
};

export type DeleteStory200 = {
  success: boolean;
};

export type ListOrders200 = {
  orders: Order[];
};

export type ListChats200 = {
  chats: Chat[];
};

export type GetChatMessages200 = {
  messages: Message[];
};