# lib/api-zod/src/generated/types/registerClientInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nixyah API - Food marketplace for Côte d'Ivoire
 * OpenAPI spec version: 0.2.0
 */

export interface RegisterClientInput {
  name: string;
  email?: string;
  phone?: string;
  password: string;
  location: string;
  preferences?: string[];
}