# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nixyah API - Food marketplace for Côte d'Ivoire
 * OpenAPI spec version: 0.2.0
 */

export interface Story {
  id: string;
  chefId: string;
  chefName: string;
  chefCoverColor: string;
  caption: string;
  dishName?: string;
  price?: number;
  emoji?: string;
  bgColor?: string;
  createdAt: string;
  expiresAt: string;
}